Add AuthContainer, move auth api from Container to AuthContainer
